app/requirements.txt
amqp==5.3.1
asttokens==3.0.0
async-timeout==5.0.1
billiard==4.2.1
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
executing==2.1.0
Flask==3.0.3
Flask-Cors==5.0.0
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
kombu==5.4.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mutagen==1.47.0
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodomex==3.21.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.8.0
six==1.17.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.6
wheel==0.45.1
yt-dlp==2024.10.22
zipp==3.20.2
SQLAlchemy==2.0.37
greenlet==3.1.1
alembic==1.14.0
psycopg2-binary==2.9.10